requirements.txt
python-pptx
Pillow
google-generativeai
python-dotenv